+++ tests/conftest.py
@@ -1,15 +1,8 @@
 from pathlib import Path
-from typing import Any
 
 import pytest
-from pytest import MonkeyPatch
 
-from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.models import (
-    AnyMergedModel,
-    ExtractedOrganization,
-    PaginatedItemsContainer,
-)
+from mex.common.models import ExtractedOrganization
 from mex.common.types import MergedPrimarySourceIdentifier
 from mex.extractors.settings import Settings
 
@@ -17,6 +10,9 @@ pytest_plugins = (
     "mex.common.testing.plugin",
     "tests.blueant.mocked_blueant",
     "tests.confluence_vvt.mocked_confluence_vvt",
+    "tests.datenkompass.mocked_backend_api_connector",
+    "tests.datenkompass.mocked_provider",
+    "tests.datenkompass.mocked_boto",
     "tests.datscha_web.mocked_datscha_web",
     "tests.drop.mocked_drop",
     "tests.grippeweb.mocked_grippeweb",
@@ -25,7 +21,6 @@ pytest_plugins = (
     "tests.ldap.mocked_ldap",
     "tests.open_data.mocked_open_data",
 )
-
 
 TEST_DATA_DIR = Path(__file__).parent / "test_data"
 
@@ -43,23 +38,3 @@ def extracted_organization_rki() -> ExtractedOrganization:
         hadPrimarySource=MergedPrimarySourceIdentifier.generate(123),
         officialName=["Robert Koch-Institut"],
     )
-
-
-@pytest.fixture
-def mocked_backend(monkeypatch: MonkeyPatch) -> None:
-    """Mock the backendAPIConnector to return dummy variables."""
-
-    def mocked_request(
-        _self: BackendApiConnector,
-        _method: str = "GET",
-        _endpoint: str | None = None,
-        _payload: Any = None,  # noqa: ANN401
-        _params: dict[str, str] | None = None,
-        **_kwargs: Any,  # noqa: ANN401
-    ) -> dict[str, Any]:
-        return PaginatedItemsContainer[AnyMergedModel](
-            total=0,
-            items=[],
-        ).model_dump()
-
-    monkeypatch.setattr(BackendApiConnector, "request", mocked_request)
